+++ packages/react/src/exampleTypes/Facet.js
@@ -1,11 +1,40 @@
 import React, { useState } from 'react'
 import _ from 'lodash/fp'
-import F from 'futil-js'
+import F from 'futil'
 import { setDisplayName } from 'recompose'
 import { observer } from 'mobx-react'
 import { Flex } from '../greyVest'
 import { contexturify } from '../utils/hoc'
 import { withTheme } from '../utils/theme'
+
+export let Cardinality = _.flow(
+  setDisplayName('Cardinality'),
+  observer
+)(({ node, tree }) => (
+  <Flex
+    className="contexture-facet-cardinality"
+    style={{ justifyContent: 'space-between' }}
+  >
+    {!!node.context.cardinality && (
+      <div>
+        Showing {_.min([node.size || 10, node.context.options.length])} of{' '}
+        {node.context.cardinality}
+      </div>
+    )}
+    {node.context.cardinality > (node.size || 10) && (
+      <div>
+        <a
+          onClick={() =>
+            tree.mutate(node.path, { size: (node.size || 10) + 10 })
+          }
+          style={{ cursor: 'pointer' }}
+        >
+          View More
+        </a>
+      </div>
+    )}
+  </Flex>
+))
 
 let SelectAll = _.flow(
   setDisplayName('SelectAll'),
@@ -116,29 +145,7 @@ let Facet = ({
         )
       })
     )(_.get('context.options', node))}
-    <Flex
-      className="contexture-facet-cardinality"
-      style={{ justifyContent: 'space-between' }}
-    >
-      {!!node.context.cardinality && (
-        <div>
-          Showing {_.min([node.size || 10, node.context.options.length])} of{' '}
-          {node.context.cardinality}
-        </div>
-      )}
-      {node.context.cardinality > (node.size || 10) && (
-        <div>
-          <a
-            onClick={() =>
-              tree.mutate(node.path, { size: (node.size || 10) + 10 })
-            }
-            style={{ cursor: 'pointer' }}
-          >
-            View More
-          </a>
-        </div>
-      )}
-    </Flex>
+    <Cardinality {...{ node, tree }} />
   </div>
 )