+++ packages/react/src/exampleTypes/utils.js
@@ -1,0 +1,6 @@
+import F from 'futil'
+
+export let selectedBinding = (...lens) => ({
+  selectedValues: F.view(...lens),
+  onChange: x => F.set(x, ...lens),
+})