+++ packages/react/src/greyVest/Grid.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import _ from 'lodash/fp'
-import F from 'futil-js'
+import F from 'futil'
 
 let formatAreas = _.flow(
   _.map(F.quote),