+++ packages/react/src/stories/DemoControls.js
@@ -1,8 +1,8 @@
-import F from 'futil-js'
+import F from 'futil'
 import React from 'react'
 import { observer } from 'mobx-react'
 import { defaultProps } from 'recompose'
-import { useLens } from '../utils/react'
+
 import { TextHighlight } from '..'
 
 export let Button = x => (
@@ -18,7 +18,7 @@ export let Button = x => (
 )
 
 let Input = observer(props => {
-  let focusing = useLens(false)
+  let focusing = React.useState(false)
   return (
     <input
       style={{
@@ -51,7 +51,7 @@ export let Highlight = x => (
 )
 
 export let DropdownItem = observer(props => {
-  let hovering = useLens(false)
+  let hovering = React.useState(false)
   return (
     <div
       style={{
@@ -67,7 +67,7 @@ export let DropdownItem = observer(props => {
 })
 
 export let PagerItem = observer(({ active, ...props }) => {
-  let hovering = useLens(false)
+  let hovering = React.useState(false)
   return (
     <span
       style={{