+++ packages/react/src/utils/react.js
@@ -1,11 +1,9 @@
 import _ from 'lodash/fp'
-import * as F from 'futil-js'
+import * as F from 'futil'
 import { useState } from 'react'
 import { mapProps } from 'recompose'
-import { expandObjectBy } from './futil'
 
-export let useLens = x => F.stateLens(useState(x))
-export let useLensObject = _.mapValues(useLens)
+export let useLensObject = _.mapValues(useState)
 
 export let getDisplayName = Component =>
   F.cascade(['displayName', 'name'], Component) || 'Unknown'
@@ -17,6 +15,6 @@ export let wrapDisplayName = (name, Component) => Wrapped => {
 
 // (k, a -> {b}) -> Component<{k: a}> -> Component<{b}>
 export let expandProp = _.flow(
-  (key, fn) => expandObjectBy(key, F.whenExists(fn)),
+  (key, fn) => F.expandObjectBy(key, F.whenExists(fn)),
   mapProps
 )